+++ angular/lib/src/common/common_directives.dart
@@ -2,7 +2,7 @@ import 'directives.dart' show CORE_DIRECTIVES;
 
 /// A collection of Angular core directives that are likely to be used in each
 /// and every Angular application. This includes core directives
-/// (e.g., NgIf and NgFor).
+/// (e.g., `NgIf` and `NgFor`).
 ///
 /// This collection can be used to quickly enumerate all the built-in directives
 /// in the `directives` property of the `@Component` annotation.
@@ -42,17 +42,18 @@ import 'directives.dart' show CORE_DIRECTIVES;
 /// One could import all the core directives at once:
 ///
 /// ```dart
-/// import 'package:angular/angular.dart' show COMMON_DIRECTIVES;
+/// import 'package:angular/angular.dart' show coreDirectives;
 /// import 'my_directives.dart' show OtherDirective;
 ///
 /// @Component(
 ///     selector: 'my-component',
 ///     templateUrl: 'my_component.html',
-///     directives: const [COMMON_DIRECTIVES, OtherDirective])
+///     directives: const [coreDirectives, OtherDirective])
 /// class MyComponent {
 ///   ...
 /// }
 /// ```
-const List<List<Type>> COMMON_DIRECTIVES = const [
-  CORE_DIRECTIVES,
-];
+const coreDirectives = CORE_DIRECTIVES;
+
+@Deprecated('Renamed to "coreDirectives"')
+const COMMON_DIRECTIVES = coreDirectives;